Move the old properties command to hidden command category.
